Footer Update
Scrolling to the bottom of the landing page makes the Footer fade in, and fade out if you scroll back up. However, if the screen is too small or too big, the footer will not appear. Fix this.

clsx library was also downloaded to allow for dynamic values in tailwind classNames